[hardware] Squash all commits to add ideal LRWait
[hardware] Add LR/SC in TCDM adapter

  Port LR/SC functionality for the TCDM adapter from snitch_cluster
  to mempool. The DMA functionality used in snitch_cluster for the
  TCDM cluster was omitted.

[software] Add riscv unit test for LR/SC

[hardware] Handle rdata_o drive conflict

[hardware] Raise flag when SC succeded

[hardware] Use absolute core_id for LRSC

The previous implementation of core_id used the tile_id which is not
absolute but relative to the group the tile is in. This lead to core_ids
that are not necessarily unique for the reservation of a Load reserve.

[hardware] Add tile_id to unique core_id

The previous configuration for unique_core_id did not provide
a sufficient width for the mempool configuration.

[hardware] Make LRSC parametrizable

[hardware] Fix RISC-V tests 2, 3 and 4

[hardware] Add single core LRSC test

A test for a successful LRSC was missing.

[hardware] Add parameter to mempool_pkg

[hardware] Remove import and clean up

[hardware] Reduce length of unique_core_id

[hardware] Fix parametrizing LR/SC

[hardware] Change bitwise to logic operators

[hardware] Remove absolute GroupId and clean-up

[hardware] Add LRWait working for low contention

[hardware] Handle contention of LRWait

The previous implementation could not handle another
instruction after a SC. Now in_valid_o is kept low to
prevent other instructions from accessing the TCDM.

In addition, the LR comes a cycle sooner and is passed
out of the TCDM before the SC response.

[hardware] Send LR response depending on rdata reg

Handle case when master is not ready for accepting new data by
applying by only sending LR response when rdata register has been
emptied.

[hardware] Handle stalling of TCDM adapter

[hardware] Make LRWait queue size parametrizable

[hardware] Handle double reservation in LRWait

[hardware] LRWait: Register address

[hardware] Remove wrong assertion

[hardware] Fix registering address for queue

[hardware] Fix double reservation LRWait